Created extra ACF field in the 'Header Settings'
In the ACF field Group, it has the name:
btn1-enable_google_analytics_tracking-in-header.

I also added a code snippet to conditionally append the onclick button event in Button Two to trigger or fire the tracking code in the <head>. This happens once the above ACF field is enabled in the Theme Settings.