Also Implemented the tracking in Header Button One
I had to further implement the tracking solution on the Header Button One, in a situation where button one is considered to be tracked too.

So, I added another field in 'Header Settings' ACF field Group, with the name: btn1-enable_google_analytics_tracking-in-header. Included the corresponding code snippet to to conditionally append the onclick button event in Button One to trigger or fire the tracking code in the <head>. This happens once the above acf field is enabled.